+++ pevents.h
@@ -21,7 +21,7 @@ namespace neosmart
     //WIN32-style functions
 	neosmart_event_t CreateEvent(bool manualReset = false, bool initialState = false);
 	int DestroyEvent(neosmart_event_t event);
-	int WaitForEvent(neosmart_event_t event, uint32_t milliseconds = -1);
+	int WaitForEvent(neosmart_event_t event, uint64_t milliseconds = -1);
 	int SetEvent(neosmart_event_t event);
 	int ResetEvent(neosmart_event_t event);
 #ifdef WFMO
